Fixed most issues mentioned, and did some (on slack) discussed stuff..
- Replaced gamemode icon (png) with the font version.
- Now including country flags automatically (NOTE: I'm using the path "/images/flags/{{ Auth::user()->country_acronym }}.png")
- Reworked the ENTIRE dropdown modal for both login and when logged in, to be based upon the BEM concept.
- Now using localization

As discussed with @nanaya on Slack, this entire has been reworked to be based upon the BEM concept:
This also means that there is now a BEM folder, and a bem-index.less file, in which each BEM-block is being included by it's own .less file.
This can be changed in the future, but for now I suppose this is okay. It's pretty much a playtest for now, doesn't really matter as it's easily moveable.