Properly set new csrf token upon login
It's handled by pageRefresh for logout.

Encryption in xsrf is meaningless because the verifier accepts both encrypted and unencrypted version as is. Except the encrypted version requires the header to be called 'X-XSRF-TOKEN'. Since they're same thing, I figured it's better to just pass them unencrypted instead of overriding jquery-ujs to pass that instead.

Or so I thought. Apparently if there's _token input, it'll always be used and never attempted to be decrypted and thus passing the token unencrypted becomes a requirement =D