Improve phpunit wrapper
Use shell script instead of php for preloading framework override before starting test. They added this ability in 7.4.

Also it fixes problem in phpunit 8 as the start script can't be run using `require` because of `strict_types` declaration being on second line.

As a bonus, it removes the appearance of shebang line in test due to how `require` works.