Generate sourcemaps for css files as well
Just setting it in webpackConfig doesn't work for css as the loaders need to be passed similar options and mix only passes it if we call `.sourceMaps()` to it.

devtool in webpackConfig is redundant as sourceMaps also set whatever it's getting as default config. Except the dev default (eval-source-map) is not working for css files so it needs to be set to the correct type.

Also explicitly set the type for production so they are always the same between environments and not rely on defaults.