faster layout on Chrome when hovering over conversation list.
The combination of a scrollbar, opacity change of the close button and
animation of the unread indicator seems to cause Chrome to do some relatively
expensive relayouts. Setting the indicator and button to position:
absolute reduces the amount of layouts getting recalculated.

Neither Firefox nor Safari appear to have this issue.