Move beatmap attributes to per-beatmap document
This will allow searching for beatmaps with specific combination of attributes.

For example, searching for beatmap with difficulty rating of 3 and drain rate of 8 using the old structure may return beatmapset which contain two different beatmaps, one with diff 3 (but maybe drain 7) and another one with drain 8 (but maybe diff 4).