Fix display of news articles with first `<img>` at top level
If the first `<img>` tag in a processed news article is at top level
in the document (via raw HTML), the entire article is hidden client-side
when attempting to remove first image by deleting the image's parent.

This fixes such cases by finding the first image by its URL, and
removing its parent only if it's a `P` element with one child.