Remove needed to be in channel when receiving messages targetted at user.
Load channel when receiving message targetted at user if it's not loaded or was removed from the store.
Since the user is still "in" the channel, they should receive the message without needed to join.
This should remove the need to account for varying latencies with the responses from php and the socket server.
Messages sent to public channels after leaving are still ignored.