Fix setHash triggering on wrong page after navigation.
Because the component of the next page mounts before the current out unmount, autorun
ends up triggering setHash if any of the observables it's watching gets updated.

It just so happens that current user playmode ends up being observed if via findDefault() through currentBeatmap()
if a beatmap hasn't explicitly been selected yet, which causes autorun to trigger when core.setCurrentUser is triggered on navigation.

A more complete fix would to be set the default beatmap when initializing `beatmapsets-show/controller` so that currentUser doesn't set sometimes observed.