Always push beatmaps to BSS queue after they become eligible for score submission
The reasoning for this change is related to the introduction of legacy
scoring attributes
(https://github.com/ppy/osu-difficulty-calculator/pull/224) which are
intended to be used by `osu-difficulty-calculator` to recalculate scores
imported from stable to the new standardised scoring scheme.

Today it was revealed that some stable scores could not be imported to
the new scoring tables due to missing legacy scoring attributes. In
particular, said failing scores were set on the following beatmap:

	https://osu.ppy.sh/beatmapsets/2004898

The current working theory which intends to support this change is that
the scoring attributes are missing for this beatmap due to
the following coincidence of circumstances:

- During back-population of scoring attributes, the beatmap in question
  was excluded due to being in pending state

- The beatmap has not been updated by the creator since the
  back-population, meaning that it wouldn't have been pushed to
  `osu-difficulty-calculator` for reprocessing

- At 20231228T130004Z the beatmap was qualified, which up until this
  change was a process involving `osu-web` only and as such would not
  queue the beatmap for BSS reprocessing.

Notably a similar sequence of events can also plausibly happen when
loving a beatmap, as it is another case wherein a beatmap enters
a state wherein scores set on it will be submitted and preserved in
the database.

The suggested fix here is to queue the beatmap for a BSS reprocess after
both of the aforementioned ranked status changes, which will - via the
submission flow - ensure that the beatmap reaches
`osu-difficulty-calculator` for reprocessing, and receive legacy scoring
attributes as a result. This was already happening for newly-ranked
beatmaps, so this is intended to simply be an extension of that.