Exclude scores without pp when displaying pack item pass indicator
This is a port of
https://github.com/ppy/osu-queue-score-statistics/pull/234 to web.

I'm not super sure this is entirely valid. For the purposes of pack
medals that was an okay workaround to apply. As long as challenge
packs are only used for medals and/or as long as no challenge packs
contain unranked/loved maps, this should hold up in web too.