Fix scores set with unranked mods showing as processing indefinitely (again)
Closes https://github.com/ppy/osu-web/issues/10911 once more.

The actual fix is the inclusion of and check for existence of the
`score_process_history` row as it is the most reliable indicator of
processing having concluded.

This also includes a drive-by inclusion of a `preserve === false` check
that won't really do anything right now, but will if / once
https://github.com/ppy/osu-queue-score-statistics/issues/141 is
resolved.

The goal here is to only show pp values on scores that contribute to the
user's total. The query determining can be found here:

	https://github.com/ppy/osu-queue-score-statistics/blob/f65dec335a914ee171d69981719d6ccb1d3f0ebe/osu.Server.Queues.ScoreStatisticsProcessor/Processors/UserTotalPerformanceProcessor.cs#L66-L76