Automatically schedule daily challenge room creation
Cron is intended to run every day at 00:05 UTC. Looking at daily
metrics, this should coincide with the tail end of the daily user peak.
The five minutes is just supposed to be a bit of a safety gap to ensure
the rooms don't run into each other.

If required, the time can be probably shifted around (userbase nadir
appears to be roughly somewhere between ~06:00 and 08:00 UTC), but if
that is to happen, then the end date of the room will also need to be
shifted forward 8 hours (currently set to 00:00 UTC of the next day).