Fix container size of beatmap title and artist
The class was originally directly applied to the search link and thus
need to be limited in width so the empty space isn't clickable.

It has since become a container and the actual link has been moved.

Filling the whole width makes selecting text actually selects the
respective line instead of whatever below/above it.