Automatically create beatmap set description topic on edit if missing
This is a solution to the problem of the new beatmap submission server
not creating a description topic for the set like the old one did.
Notably, this solution implies several shortcuts:

- If the main metadata of the set (artist, title) don't change, the
  forum thread will not be renamed
- The forum thread will be posted in a hardcoded forum and never moved,
  while old BSS would move the thread between forums 6 and 10 depending
  on whether the beatmap is pending or WIP. Not doing this saves a whole
  bunch of busy work.
- There is no actual body in the forum thread OP, contrary to what old
  submission did.

The aforementioned simplifications were made with the assumption that
subforums 6 and 10 on production were going to be hidden away from users
due to not containing any actual content that anyone would ever want to
see. (Same thing applies to forum 19, which is graveyard.)

The *one and only goal* here is that descriptions should be editable for
maps submitted using new BSS. Thus I am taking pretty much the same
shortcuts that places like the profile page already seem to be.