Replace compound primary key on `beatmapset_version_files` with synthetic one
The primary key was on `(file_id, version_id)`, which looks fine in
theory, but in practice I hit a case where I got primary key uniqueness
violations, and realised that this setup fails if a beatmap set has two
or more files in the package with identical contents but different
filenames.

This is the case, since `beatmapset_files` has a column for the files'
SHA256, with a uniqueness constraint, so two files with identical
content will refer back to the same `beatmapset_files` row.

I considered adding `filename` to the PK but decided against it since
it's probably slow and at this point attempting to conjure a natural key
is clearly a lost cause.